requirements.txt
python-telegram-bot<20
requests
pymysql
prettytable
pytz
python-dateutil
pycryptodome